// LocadoraCrud/LocadoraApp/src/app/models/Locadora.ts
export class Locadora {
    id: number = 0;
    nomeFantasia: string = '';
    razaoSocial: string = '';
    cnpj: string = '';
    email: string = '';
    telefone: string = '';
    cep: string = '';
    rua: string = '';
    numero: string = '';
    bairro: string = '';
    estado: string = '';
    cidade: string = '';
  }
  